+++ exporters/prometheus/prometheus.go
@@ -131,9 +131,9 @@ func (e *Exporter) Controller() *controller.Controller {
 	return e.controller
 }
 
-// ExportKindFor implements ExportKindSelector.
-func (e *Exporter) ExportKindFor(desc *metric.Descriptor, kind aggregation.Kind) export.ExportKind {
-	return export.CumulativeExportKindSelector().ExportKindFor(desc, kind)
+// TemporalityFor implements TemporalitySelector.
+func (e *Exporter) TemporalityFor(desc *metric.Descriptor, kind aggregation.Kind) export.Temporality {
+	return export.CumulativeTemporalitySelector().TemporalityFor(desc, kind)
 }
 
 // ServeHTTP implements http.Handler.