+++ sdk/export/metric/metric.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:generate stringer -type=ExportKind
+//go:generate stringer -type=Temporality
 
 package metric // import "go.opentelemetry.io/otel/sdk/export/metric"
 
@@ -220,20 +220,20 @@ type Exporter interface {
 	// Processor that just completed collection.
 	Export(ctx context.Context, resource *resource.Resource, reader InstrumentationLibraryReader) error
 
-	// ExportKindSelector is an interface used by the Processor
+	// TemporalitySelector is an interface used by the Processor
 	// in deciding whether to compute Delta or Cumulative
 	// Aggregations when passing Records to this Exporter.
-	ExportKindSelector
+	TemporalitySelector
 }
 
-// ExportKindSelector is a sub-interface of Exporter used to indicate
+// TemporalitySelector is a sub-interface of Exporter used to indicate
 // whether the Processor should compute Delta or Cumulative
 // Aggregations.
-type ExportKindSelector interface {
-	// ExportKindFor should return the correct ExportKind that
+type TemporalitySelector interface {
+	// TemporalityFor should return the correct Temporality that
 	// should be used when exporting data for the given metric
 	// instrument and Aggregator kind.
-	ExportKindFor(descriptor *metric.Descriptor, aggregatorKind aggregation.Kind) ExportKind
+	TemporalityFor(descriptor *metric.Descriptor, aggregatorKind aggregation.Kind) Temporality
 }
 
 // InstrumentationLibraryReader is an interface for exporters to iterate
@@ -255,7 +255,7 @@ type Reader interface {
 	// period. Each aggregated checkpoint returned by the
 	// function parameter may return an error.
 	//
-	// The ExportKindSelector argument is used to determine
+	// The TemporalitySelector argument is used to determine
 	// whether the Record is computed using Delta or Cumulative
 	// aggregation.
 	//
@@ -263,7 +263,7 @@ type Reader interface {
 	// expected from the Meter implementation. Any other kind
 	// of error will immediately halt ForEach and return
 	// the error to the caller.
-	ForEach(kindSelector ExportKindSelector, recordFunc func(Record) error) error
+	ForEach(kindSelector TemporalitySelector, recordFunc func(Record) error) error
 
 	// Locker supports locking the checkpoint set.  Collection
 	// into the checkpoint set cannot take place (in case of a
@@ -366,38 +366,38 @@ func (r Record) EndTime() time.Time {
 	return r.end
 }
 
-// ExportKind indicates the kind of data exported by an exporter.
+// Temporality indicates the kind of data exported by an exporter.
 // These bits may be OR-d together when multiple exporters are in use.
-type ExportKind int
+type Temporality int
 
 const (
-	// CumulativeExportKind indicates that an Exporter expects a
+	// CumulativeTemporality indicates that an Exporter expects a
 	// Cumulative Aggregation.
-	CumulativeExportKind ExportKind = 1
+	CumulativeTemporality Temporality = 1
 
-	// DeltaExportKind indicates that an Exporter expects a
+	// DeltaTemporality indicates that an Exporter expects a
 	// Delta Aggregation.
-	DeltaExportKind ExportKind = 2
+	DeltaTemporality Temporality = 2
 )
 
 // Includes tests whether `kind` includes a specific kind of
 // exporter.
-func (kind ExportKind) Includes(has ExportKind) bool {
+func (kind Temporality) Includes(has Temporality) bool {
 	return kind&has != 0
 }
 
 // MemoryRequired returns whether an exporter of this kind requires
 // memory to export correctly.
-func (kind ExportKind) MemoryRequired(mkind sdkapi.InstrumentKind) bool {
+func (kind Temporality) MemoryRequired(mkind sdkapi.InstrumentKind) bool {
 	switch mkind {
 	case sdkapi.HistogramInstrumentKind, sdkapi.GaugeObserverInstrumentKind,
 		sdkapi.CounterInstrumentKind, sdkapi.UpDownCounterInstrumentKind:
 		// Delta-oriented instruments:
-		return kind.Includes(CumulativeExportKind)
+		return kind.Includes(CumulativeTemporality)
 
 	case sdkapi.CounterObserverInstrumentKind, sdkapi.UpDownCounterObserverInstrumentKind:
 		// Cumulative-oriented instruments:
-		return kind.Includes(DeltaExportKind)
+		return kind.Includes(DeltaTemporality)
 	}
 	// Something unexpected is happening--we could panic.  This
 	// will become an error when the exporter tries to access a
@@ -406,49 +406,49 @@ func (kind ExportKind) MemoryRequired(mkind sdkapi.InstrumentKind) bool {
 }
 
 type (
-	constantExportKindSelector  ExportKind
-	statelessExportKindSelector struct{}
+	constantTemporalitySelector  Temporality
+	statelessTemporalitySelector struct{}
 )
 
 var (
-	_ ExportKindSelector = constantExportKindSelector(0)
-	_ ExportKindSelector = statelessExportKindSelector{}
+	_ TemporalitySelector = constantTemporalitySelector(0)
+	_ TemporalitySelector = statelessTemporalitySelector{}
 )
 
-// ConstantExportKindSelector returns an ExportKindSelector that returns
-// a constant ExportKind, one that is either always cumulative or always delta.
-func ConstantExportKindSelector(kind ExportKind) ExportKindSelector {
-	return constantExportKindSelector(kind)
+// ConstantTemporalitySelector returns an TemporalitySelector that returns
+// a constant Temporality, one that is either always cumulative or always delta.
+func ConstantTemporalitySelector(kind Temporality) TemporalitySelector {
+	return constantTemporalitySelector(kind)
 }
 
-// CumulativeExportKindSelector returns an ExportKindSelector that
-// always returns CumulativeExportKind.
-func CumulativeExportKindSelector() ExportKindSelector {
-	return ConstantExportKindSelector(CumulativeExportKind)
+// CumulativeTemporalitySelector returns an TemporalitySelector that
+// always returns CumulativeTemporality.
+func CumulativeTemporalitySelector() TemporalitySelector {
+	return ConstantTemporalitySelector(CumulativeTemporality)
 }
 
-// DeltaExportKindSelector returns an ExportKindSelector that
-// always returns DeltaExportKind.
-func DeltaExportKindSelector() ExportKindSelector {
-	return ConstantExportKindSelector(DeltaExportKind)
+// DeltaTemporalitySelector returns an TemporalitySelector that
+// always returns DeltaTemporality.
+func DeltaTemporalitySelector() TemporalitySelector {
+	return ConstantTemporalitySelector(DeltaTemporality)
 }
 
-// StatelessExportKindSelector returns an ExportKindSelector that
-// always returns the ExportKind that avoids long-term memory
+// StatelessTemporalitySelector returns an TemporalitySelector that
+// always returns the Temporality that avoids long-term memory
 // requirements.
-func StatelessExportKindSelector() ExportKindSelector {
-	return statelessExportKindSelector{}
+func StatelessTemporalitySelector() TemporalitySelector {
+	return statelessTemporalitySelector{}
 }
 
-// ExportKindFor implements ExportKindSelector.
-func (c constantExportKindSelector) ExportKindFor(_ *metric.Descriptor, _ aggregation.Kind) ExportKind {
-	return ExportKind(c)
+// TemporalityFor implements TemporalitySelector.
+func (c constantTemporalitySelector) TemporalityFor(_ *metric.Descriptor, _ aggregation.Kind) Temporality {
+	return Temporality(c)
 }
 
-// ExportKindFor implements ExportKindSelector.
-func (s statelessExportKindSelector) ExportKindFor(desc *metric.Descriptor, kind aggregation.Kind) ExportKind {
+// TemporalityFor implements TemporalitySelector.
+func (s statelessTemporalitySelector) TemporalityFor(desc *metric.Descriptor, kind aggregation.Kind) Temporality {
 	if kind == aggregation.SumKind && desc.InstrumentKind().PrecomputedSum() {
-		return CumulativeExportKind
+		return CumulativeTemporality
 	}
-	return DeltaExportKind
+	return DeltaTemporality
 }