+++ sdk/metric/controller/basic/pull_test.go
@@ -36,7 +36,7 @@ func TestPullNoCollect(t *testing.T) {
 	puller := controller.New(
 		processor.NewFactory(
 			processortest.AggregatorSelector(),
-			export.CumulativeExportKindSelector(),
+			export.CumulativeTemporalitySelector(),
 			processor.WithMemory(true),
 		),
 		controller.WithCollectPeriod(0),
@@ -51,7 +51,7 @@ func TestPullNoCollect(t *testing.T) {
 
 	require.NoError(t, puller.Collect(ctx))
 	records := processortest.NewOutput(attribute.DefaultEncoder())
-	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeExportKindSelector(), records.AddInstrumentationLibraryRecord))
+	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeTemporalitySelector(), records.AddInstrumentationLibraryRecord))
 
 	require.EqualValues(t, map[string]float64{
 		"counter.sum/A=B/": 10,
@@ -61,7 +61,7 @@ func TestPullNoCollect(t *testing.T) {
 
 	require.NoError(t, puller.Collect(ctx))
 	records = processortest.NewOutput(attribute.DefaultEncoder())
-	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeExportKindSelector(), records.AddInstrumentationLibraryRecord))
+	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeTemporalitySelector(), records.AddInstrumentationLibraryRecord))
 
 	require.EqualValues(t, map[string]float64{
 		"counter.sum/A=B/": 20,
@@ -72,7 +72,7 @@ func TestPullWithCollect(t *testing.T) {
 	puller := controller.New(
 		processor.NewFactory(
 			processortest.AggregatorSelector(),
-			export.CumulativeExportKindSelector(),
+			export.CumulativeTemporalitySelector(),
 			processor.WithMemory(true),
 		),
 		controller.WithCollectPeriod(time.Second),
@@ -89,7 +89,7 @@ func TestPullWithCollect(t *testing.T) {
 
 	require.NoError(t, puller.Collect(ctx))
 	records := processortest.NewOutput(attribute.DefaultEncoder())
-	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeExportKindSelector(), records.AddInstrumentationLibraryRecord))
+	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeTemporalitySelector(), records.AddInstrumentationLibraryRecord))
 
 	require.EqualValues(t, map[string]float64{
 		"counter.sum/A=B/": 10,
@@ -100,7 +100,7 @@ func TestPullWithCollect(t *testing.T) {
 	// Cached value!
 	require.NoError(t, puller.Collect(ctx))
 	records = processortest.NewOutput(attribute.DefaultEncoder())
-	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeExportKindSelector(), records.AddInstrumentationLibraryRecord))
+	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeTemporalitySelector(), records.AddInstrumentationLibraryRecord))
 
 	require.EqualValues(t, map[string]float64{
 		"counter.sum/A=B/": 10,
@@ -112,7 +112,7 @@ func TestPullWithCollect(t *testing.T) {
 	// Re-computed value!
 	require.NoError(t, puller.Collect(ctx))
 	records = processortest.NewOutput(attribute.DefaultEncoder())
-	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeExportKindSelector(), records.AddInstrumentationLibraryRecord))
+	require.NoError(t, controllertest.ReadAll(puller, export.CumulativeTemporalitySelector(), records.AddInstrumentationLibraryRecord))
 
 	require.EqualValues(t, map[string]float64{
 		"counter.sum/A=B/": 20,