+++ sdk/metric/controller/controllertest/test.go
@@ -64,7 +64,7 @@ func (t MockTicker) C() <-chan time.Time {
 // metric).
 func ReadAll(
 	reader export.InstrumentationLibraryReader,
-	kind export.ExportKindSelector,
+	kind export.TemporalitySelector,
 	apply func(instrumentation.Library, export.Record) error,
 ) error {
 	return reader.ForEach(func(library instrumentation.Library, reader export.Reader) error {